backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.4.0"
description = ""
authors = ["seungju"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.1"
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"
httpx = "^0.27.2"
python-socketio = "^5.11.4"
eventlet = "^0.38.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
